# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tripzant API Documentation",
        "description": "API endpoints for Tripzant Mobile App",
        "contact": {
            "email": "support@tripzant.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://tripnstays.com/api/v1",
            "description": "Staging/Live Server"
        }
    ],
    "paths": {
...
        "/payment/initiate": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Initialize Payment (Mobile App)",
                "description": "Create a payment session for Stripe or MPGS (Commercial Bank)",
                "operationId": "c03375ef12159bfa368107598caa599e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "total_amount",
                                    "gateway",
                                    "travelers"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "flight"
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "gateway": {
                                        "type": "string",
                                        "enum": [
                                            "stripe",
                                            "mpgs"
                                        ]
                                    },
                                    "travelers": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "item_data": {
                                        "type": "object"
                                    },
                                    "reference": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment session created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/payment/success": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Confirm Payment (Mobile App)",
                "description": "Verify payment status and confirm the booking record",
                "operationId": "a54e7ddeaa3726396dfc918148e6f4ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reference"
                                ],
                                "properties": {
                                    "reference": {
                                        "type": "string",
                                        "example": "API-ABC123"
                                    },
                                    "transaction_id": {
                                        "type": "string"
                                    },
                                    "gateway": {
                                        "type": "string",
                                        "enum": [
                                            "stripe",
                                            "mpgs"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking confirmed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}